feat: refactor Card components to use CSS modules and improve styling
