+++ components/Atomic/Card/CardButton.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import style from './style.module.css'
+
+const CardButton = () => {
+  return <button className={style.cardButton}>Read More</button>
+}
+
+export default CardButton
