+++ components/Atomic/Card/CardData.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import style from './style.module.css'
+const CardData = () => {
+  return <p className={style.cardData}>12/2024</p>
+}
+
+export default CardData
