+++ components/Atomic/Card/CardImage.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
+import style from './style.module.css'
 
 const CardImage = () => {
   return (
-    <div className="card-right">
+    <div className={style.cardImage}>
       <img
         src="https://cdn.mos.cms.futurecdn.net/dP3N4qnEZ4tCTCLq59iysd.jpg"
         alt="Card Image"
