+++ components/Atomic/Card/CardTitle.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import style from './style.module.css'
+const CardTitle = () => {
+  return <h1 className={style.title}>Lorem Ipsum</h1>
+}
+
+export default CardTitle
