--- components/Atomic/CardButton/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const CardButton = () => {
-  return <button className="card-button">Read More</button>
-}
-
-export default CardButton
