--- components/Atomic/CardData/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const CardData = () => {
-  return <p className="card-data">12/2024</p>
-}
-
-export default CardData
