--- components/Atomic/CardTags/index.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-const CardTags = () => {
-  return (
-    <div className="card-tags">
-      <span className="tag">Tag1</span>
-      <span className="tag">Tag2</span>
-      <span className="tag">Tag3</span>
-    </div>
-  )
-}
-
-export default CardTags
