--- components/Atomic/CardTitle/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const CardTitle = () => {
-  return <h1 className="card-title">Lorem Ipsum</h1>
-}
-
-export default CardTitle
